Re-enable rustfmt of generated bindings
Because recent versions of bindgen generate their bindings on one line,
because of bindgen's use of u128 and rust's warnings of its unsafety
FFI-wise, and because of https://github.com/rust-lang/rust/issues/62999,
combined, building the crate generates massive logs... which exceed
travis's maximum log length.

It was disabled in #13 because back then rustfmt had some problems with
the bindgen output. Now that rustfmt is part of rust, these kind of
problems presumably should happen less if at all.