Fix AudioUnit linking on iOS
On iOS, the AudioUnit framework is not (and newer were) linkable, instead one should link to AudioToolbox.

Source: https://stackoverflow.com/a/14120936

Verified by looking at simulator runtimes (this is what I could find) for version 9.3 and 10.3, on both of these AudioUnit does not contain a dylib (unlike most other frameworks). Also, in the SDK it's missing an AudioUnit.tbd file in:
/Applications/Xcode.app/Contents/Developer/Platforms/iPhoneOS.platform/Developer/SDKs/iPhoneOS.sdk/System/Library/Frameworks/AudioUnit.framework